# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TES Discord Bot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  online: boolean;
  tag: string | null;
  uptime: number | null;
  guildCount: number;
  prefix: string;
}

export interface UserBalance {
  userId: string;
  username: string;
  balance: number;
}

export interface BotStats {
  totalUsers: number;
  totalTokensInCirculation: number;
  totalOrders: number;
  pendingOrders: number;
  topBalances: UserBalance[];
}

export type UserServices = { [key: string]: boolean };

export interface InventoryItem {
  id: string;
  name: string;
  acquiredAt: string;
}

export interface User {
  id: number;
  discordId: string;
  username: string;
  balance: number;
  inventory: InventoryItem[];
  services: UserServices;
  createdAt: string;
  updatedAt: string;
}

export type UpdateBalanceBodyOperation =
  (typeof UpdateBalanceBodyOperation)[keyof typeof UpdateBalanceBodyOperation];

export const UpdateBalanceBodyOperation = {
  add: "add",
  subtract: "subtract",
  set: "set",
} as const;

export interface UpdateBalanceBody {
  amount: number;
  operation: UpdateBalanceBodyOperation;
}

export type ShopItemType = (typeof ShopItemType)[keyof typeof ShopItemType];

export const ShopItemType = {
  item: "item",
  service: "service",
} as const;

export interface ShopItem {
  id: string;
  label: string;
  price: number;
  type: ShopItemType;
  description: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId: string;
  username: string;
  itemId: string;
  itemLabel: string;
  price: number;
  status: OrderStatus;
  createdAt: string;
  updatedAt: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type ActivityEventType =
  (typeof ActivityEventType)[keyof typeof ActivityEventType];

export const ActivityEventType = {
  purchase: "purchase",
  token_add: "token_add",
  order_placed: "order_placed",
  order_completed: "order_completed",
} as const;

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  userId: string;
  username: string;
  description: string;
  amount?: number | null;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetUsersParams = {
  /**
   * Search by Discord user ID or username
   */
  search?: string;
  /**
   * Field to sort by
   */
  sortBy?: GetUsersSortBy;
  /**
   * Sort order
   */
  order?: GetUsersOrder;
};

export type GetUsersSortBy =
  (typeof GetUsersSortBy)[keyof typeof GetUsersSortBy];

export const GetUsersSortBy = {
  balance: "balance",
  username: "username",
  createdAt: "createdAt",
} as const;

export type GetUsersOrder = (typeof GetUsersOrder)[keyof typeof GetUsersOrder];

export const GetUsersOrder = {
  asc: "asc",
  desc: "desc",
} as const;

export type GetOrdersParams = {
  /**
   * Filter by order status
   */
  status?: GetOrdersStatus;
};

export type GetOrdersStatus =
  (typeof GetOrdersStatus)[keyof typeof GetOrdersStatus];

export const GetOrdersStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;