# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TES Discord Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns the current status of the Discord bot
 * @summary Get bot status
 */
export const GetBotStatusResponse = zod.object({
  online: zod.boolean(),
  tag: zod.string().nullable(),
  uptime: zod.number().nullable(),
  guildCount: zod.number(),
  prefix: zod.string(),
});

/**
 * Returns aggregate statistics about the bot
 * @summary Get bot statistics
 */
export const GetBotStatsResponse = zod.object({
  totalUsers: zod.number(),
  totalTokensInCirculation: zod.number(),
  totalOrders: zod.number(),
  pendingOrders: zod.number(),
  topBalances: zod.array(
    zod.object({
      userId: zod.string(),
      username: zod.string(),
      balance: zod.number(),
    }),
  ),
});

/**
 * Returns all registered users with their balances and inventories
 * @summary Get all users
 */
export const GetUsersQueryParams = zod.object({
  search: zod.coerce
    .string()
    .optional()
    .describe("Search by Discord user ID or username"),
  sortBy: zod
    .enum(["balance", "username", "createdAt"])
    .optional()
    .describe("Field to sort by"),
  order: zod.enum(["asc", "desc"]).optional().describe("Sort order"),
});

export const GetUsersResponseItem = zod.object({
  id: zod.number(),
  discordId: zod.string(),
  username: zod.string(),
  balance: zod.number(),
  inventory: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      acquiredAt: zod.coerce.date(),
    }),
  ),
  services: zod.record(zod.string(), zod.boolean()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetUsersResponse = zod.array(GetUsersResponseItem);

/**
 * Returns a specific user's data by Discord ID
 * @summary Get a specific user
 */
export const GetUserParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserResponse = zod.object({
  id: zod.number(),
  discordId: zod.string(),
  username: zod.string(),
  balance: zod.number(),
  inventory: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      acquiredAt: zod.coerce.date(),
    }),
  ),
  services: zod.record(zod.string(), zod.boolean()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * Admin endpoint to add or set tokens for a user
 * @summary Update user balance
 */
export const UpdateUserBalanceParams = zod.object({
  userId: zod.coerce.string(),
});

export const UpdateUserBalanceBody = zod.object({
  amount: zod.number(),
  operation: zod.enum(["add", "subtract", "set"]),
});

export const UpdateUserBalanceResponse = zod.object({
  id: zod.number(),
  discordId: zod.string(),
  username: zod.string(),
  balance: zod.number(),
  inventory: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      acquiredAt: zod.coerce.date(),
    }),
  ),
  services: zod.record(zod.string(), zod.boolean()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * Returns all items available in the TES shop
 * @summary Get shop items
 */
export const GetShopItemsResponseItem = zod.object({
  id: zod.string(),
  label: zod.string(),
  price: zod.number(),
  type: zod.enum(["item", "service"]),
  description: zod.string(),
});
export const GetShopItemsResponse = zod.array(GetShopItemsResponseItem);

/**
 * Returns all service orders placed in the shop
 * @summary Get all orders
 */
export const GetOrdersQueryParams = zod.object({
  status: zod
    .enum(["pending", "completed", "cancelled"])
    .optional()
    .describe("Filter by order status"),
});

export const GetOrdersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  username: zod.string(),
  itemId: zod.string(),
  itemLabel: zod.string(),
  price: zod.number(),
  status: zod.enum(["pending", "completed", "cancelled"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetOrdersResponse = zod.array(GetOrdersResponseItem);

/**
 * Update the status of a service order
 * @summary Update order status
 */
export const UpdateOrderStatusParams = zod.object({
  orderId: zod.coerce.number(),
});

export const UpdateOrderStatusBody = zod.object({
  status: zod.enum(["pending", "completed", "cancelled"]),
});

export const UpdateOrderStatusResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  username: zod.string(),
  itemId: zod.string(),
  itemLabel: zod.string(),
  price: zod.number(),
  status: zod.enum(["pending", "completed", "cancelled"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * Returns recent shop transactions and events
 * @summary Get recent shop activity
 */
export const GetShopActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["purchase", "token_add", "order_placed", "order_completed"]),
  userId: zod.string(),
  username: zod.string(),
  description: zod.string(),
  amount: zod.number().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetShopActivityResponse = zod.array(GetShopActivityResponseItem);