# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TES Discord Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventType } from "./activityEventType";

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  userId: string;
  username: string;
  description: string;
  amount?: number | null;
  createdAt: Date;
}