# lib/api-zod/src/generated/types/botStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TES Discord Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserBalance } from "./userBalance";

export interface BotStats {
  totalUsers: number;
  totalTokensInCirculation: number;
  totalOrders: number;
  pendingOrders: number;
  topBalances: UserBalance[];
}