# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TES Discord Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  userId: string;
  username: string;
  itemId: string;
  itemLabel: string;
  price: number;
  status: OrderStatus;
  createdAt: Date;
  updatedAt: Date;
}