# lib/api-zod/src/generated/types/updateBalanceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TES Discord Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBalanceBodyOperation } from "./updateBalanceBodyOperation";

export interface UpdateBalanceBody {
  amount: number;
  operation: UpdateBalanceBodyOperation;
}