# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TES Discord Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItem } from "./inventoryItem";
import type { UserServices } from "./userServices";

export interface User {
  id: number;
  discordId: string;
  username: string;
  balance: number;
  inventory: InventoryItem[];
  services: UserServices;
  createdAt: Date;
  updatedAt: Date;
}