# lib/api-zod/src/generated/types/userBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TES Discord Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserBalance {
  userId: string;
  username: string;
  balance: number;
}